12/25質問用プッシュ
ProductControllerにて、@AuthenticationPrincipalを利用してIDを取得しようとしているが、取得までの道筋が分からない

+++ slshop-frontend/src/main/java/com/slshop/cart/CartService.java
@@ -20,7 +20,7 @@ public class CartService {
 		return cartMapper.findAll();
 	}
 
-	public void insert(Integer customerId,Integer productId, int quantity) {
+	public void insert(Long customerId,Integer productId, int quantity) {
 		this.cartMapper.insert(customerId,productId, quantity);
 	};
 }
